+++ search-films/src/hooks/localstorage.hook/useLocalStorage.d.ts
@@ -1,0 +1,4 @@
+export interface ProfileElement {
+    name: string;
+    isLogined: boolean
+};